+++ include/atheme/scrypt.h
@@ -1,0 +1,19 @@
+/*
+ * SPDX-License-Identifier: ISC
+ * SPDX-URL: https://spdx.org/licenses/ISC.html
+ *
+ * Copyright (C) 2020 Atheme Development Group (https://atheme.github.io/)
+ */
+
+#ifndef ATHEME_INC_SCRYPT_H
+#define ATHEME_INC_SCRYPT_H 1
+
+#define ATHEME_SCRYPT_MEMLIMIT_MIN  14U
+#define ATHEME_SCRYPT_MEMLIMIT_DEF  14U
+#define ATHEME_SCRYPT_MEMLIMIT_MAX  26U
+
+#define ATHEME_SCRYPT_OPSLIMIT_MIN  32768U
+#define ATHEME_SCRYPT_OPSLIMIT_DEF  524288U
+#define ATHEME_SCRYPT_OPSLIMIT_MAX  4294967295U
+
+#endif /* !ATHEME_INC_SCRYPT_H */